restrict deleting author if it have books

+++ BookStore2/BookStore.BL/Services/AuthorService.cs
@@ -11,25 +11,24 @@ namespace BookStore.BL.Services
 {
     public class AuthorService : IAuthorService
     {
-        public readonly IAuthorRepository _authorRepository;
-
+        public readonly IAuthorRepository _authorRepository; 
+        private readonly IBookRepository _bookRepository;
         private readonly IMapper _mapper;
-
         private readonly ILogger<AuthorService> _logger;
 
-        public AuthorService(IAuthorRepository authorRepository, IMapper mapper, ILogger<AuthorService> logger)
+        public AuthorService(IAuthorRepository authorRepository, IMapper mapper, ILogger<AuthorService> logger, IBookRepository bookRepository)
         {
             _authorRepository = authorRepository;
             _mapper = mapper;
             _logger = logger;
+            _bookRepository = bookRepository;
         }
 
-        public AddAuthorResponse AddAuthor(AddAuthorRequest authorRequest)
+        public async Task<AddAuthorResponse> AddAuthor(AddAuthorRequest authorRequest)
         {
             try
             {
-
-                var auth = _authorRepository.GetAuthorByName(authorRequest.Name);
+                var auth =  await _authorRepository.GetAuthorByName(authorRequest.Name);
 
                 if (auth != null)
 
@@ -40,7 +39,7 @@ namespace BookStore.BL.Services
                     };
 
                 var author = _mapper.Map<Author>(authorRequest);
-                var result = _authorRepository.AddAuthor(author);
+                var result = await _authorRepository.AddAuthor(author);
 
                 return new AddAuthorResponse()
                 {
@@ -55,9 +54,9 @@ namespace BookStore.BL.Services
             }
         }
 
-        public UpdateAuthorResponse UpdateAuthor(UpdateAuthorRequest authorRequest)
+        public async Task<UpdateAuthorResponse> UpdateAuthor(UpdateAuthorRequest authorRequest)
         {
-            var auth = _authorRepository.GetAuthorByName(authorRequest.Name);
+            var auth = await _authorRepository.GetAuthorByName(authorRequest.Name);
 
             if (auth == null)
                 return new UpdateAuthorResponse()
@@ -67,7 +66,7 @@ namespace BookStore.BL.Services
                 };
 
             var author = _mapper.Map<Author>(authorRequest);
-            var result = _authorRepository.UpdateAuthor(author);
+            var result = await _authorRepository.UpdateAuthor(author);
 
             return new UpdateAuthorResponse()
             {
@@ -76,28 +75,35 @@ namespace BookStore.BL.Services
             };
         }
 
-        public Author? DeleteAutor(int authorId)
+        public async Task<Author?> DeleteAutor(int authorId)
         {
             var author = _authorRepository.GetById(authorId);
 
+            var isAuthorHaveBook = await _bookRepository.GetBookByAuthorId(authorId);
+
+            if (isAuthorHaveBook != null)
+            {
+                throw new Exception("The author can't be deleted");
+            }
+
             _authorRepository.DeleteAutor(authorId);
 
-            return author;
+            return await author;
         }
 
-        public IEnumerable<Author> GetAllAuthors()
+        public async Task<IEnumerable<Author>> GetAllAuthors()
         {
-            return _authorRepository.GetAllAuthors();
+            return await _authorRepository.GetAllAuthors();
         }
 
-        public Author? GetAuthorByName(string name)
+        public async Task<Author?> GetAuthorByName(string name)
         {
-            return _authorRepository.GetAuthorByName(name);
+            return await _authorRepository.GetAuthorByName(name);
         }
 
-        public Author? GetById(int id)
+        public async Task<Author?> GetById(int id)
         {
-            return _authorRepository.GetById(id);
+            return await _authorRepository.GetById(id);
         }
